Relocate to `lib/kramdown/parser/gfm/options.rb`
