+++ lib/kramdown/parser/gfm/options.rb
